Chore: Move HomePage into home folder
