+++ lib/features/home/pages/HomePage.dart
@@ -1,11 +1,11 @@
-import 'package:decentproof/pages/homepage/uiblocks/OptionsMenu.dart';
-import 'package:decentproof/pages/homepage/uiblocks/SettingsButton.dart';
-import 'package:decentproof/pages/homepage/uiblocks/SupportButton.dart';
 import 'package:flutter/material.dart';
 
-import '../../shared/Integrety/ApiKeyManager.dart';
-import '../../shared/Integrety/AppCheckWrapper.dart';
-import '../../shared/Integrety/SecureStorageWrapper.dart';
+import '../../../shared/Integrety/ApiKeyManager.dart';
+import '../../../shared/Integrety/AppCheckWrapper.dart';
+import '../../../shared/Integrety/SecureStorageWrapper.dart';
+import '../uiblocks/OptionsMenu.dart';
+import '../uiblocks/SettingsButton.dart';
+import '../uiblocks/SupportButton.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({Key? key}) : super(key: key);
